requirements.txt
# FastAPI ve Web Framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
jinja2>=3.1.0,<3.2.0

# Veritabanı
sqlalchemy>=2.0.20,<2.1.0
alembic>=1.12.0,<1.13.0
psycopg2-binary>=2.9.8,<2.10.0
redis>=5.0.0,<5.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-decouple>=3.8.0,<3.9.0

# Pydantic & Validation
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.1.0,<2.2.0

# AI & ML Libraries
google-generativeai>=0.6.0,<0.7.0
google-ai-generativelanguage>=0.6.4,<0.7.0
langchain>=0.1.10,<0.2.0
langchain-community>=0.0.16,<0.1.0
langchain-google-genai>=0.0.6,<0.1.0
langgraph>=0.0.20,<0.1.0
sentence-transformers>=2.2.2,<2.3.0
huggingface_hub>=0.16.0,<0.17.0
pinecone>=5.0.0,<6.0.0

# Data Analysis & Time Series
pandas>=2.1.4,<2.2.0
numpy>=1.24.4,<1.25.0
prophet>=1.1.5,<1.2.0
scikit-learn>=1.3.2,<1.4.0

# Web Scraping & APIs
requests>=2.31.0,<2.32.0
beautifulsoup4>=4.12.2,<4.13.0
pytrends>=4.9.2,<4.10.0
serpapi>=0.1.5,<0.2.0

# HTTP Client
httpx>=0.25.2,<0.26.0
aiohttp>=3.9.1,<3.10.0

# PDF Processing
PyPDF2>=3.0.1,<3.1.0
pdfplumber>=0.10.3,<0.11.0

# Utilities
python-dotenv>=1.0.0,<1.1.0
loguru>=0.7.2,<0.8.0
tenacity>=8.2.3,<8.3.0

# Development & Testing
pytest>=7.4.3,<7.5.0
pytest-asyncio>=0.21.1,<0.22.0
black>=23.11.0,<24.0.0
isort>=5.12.0,<5.13.0

# MCP & Agent Communication
websockets>=12.0,<12.1.0
asyncio-mqtt>=0.13.0,<0.14.0

# Performance & Monitoring
prometheus-client>=0.19.0,<0.20.0 